SKIN_DISEASE_FLASk/requirements.txt
Flask==2.3.3
tensorflow==2.13.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
Werkzeug==2.3.7
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2